refactor: move team card to ui components
